P0012R1 Make exception specifications be part of the type system, version 5
The change to [except.spec] 15.4p1 could not be applied, but the current
wording looks correct.  Other changes did not apply cleanly due to minor
wording changes, but were able to be resolved editorially.
Editorially fixed formatting errors in wording, reformatted text for
consistency and readability, and fixed use of macros.