P0083R3 Splicing Maps and Sets (Revision 5)
I put the new sub-clause [container.node] as 23.1.1 not as 23.X since it
didn't seem to belong at that level. Another option would be to put it
as 23.2.4 before [associative.reqmts] and [unord.req].

In Table 111 the a_uniq.insert(nh) and a.merge(a2) rows overflow the
final column.