Reposition declval's Example
Based on its current position, this Example can be interpreted either (1) as its own, but unnumbered, paragraph, or (2) as attached (with an unfortunate intervening line break) to the preceding Remarks paragraph but whose provision the Example does not really clarify.  It seems to make much more sense to attach the example, which is rather general in nature, to the introductory paragraph of this subclause.  I s no reason to give the Example its own paragraph number, but would not object if others feel strongly to the contrary.